# internal/api/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Motus API
  version: "0.1.0"
  description: GPS tracking and fleet management API
servers:
  - url: /
    description: This server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie set after login
    bearerAuth:
      type: http
      scheme: bearer
      description: API key token
    xAuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: iOS PWA fallback — pass session_id value here
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "resource not found"
    Device:
      type: object
      required: [id, uniqueId, name, status, disabled, attributes, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        uniqueId: { type: string }
        name: { type: string }
        protocol: { type: string }
        status: { type: string }
        speedLimit: { type: number, format: double, nullable: true }
        lastUpdate: { type: string, format: date-time, nullable: true }
        positionId: { type: integer, format: int64, nullable: true }
        groupId: { type: integer, format: int64, nullable: true }
        phone: { type: string, nullable: true }
        model: { type: string, nullable: true }
        contact: { type: string, nullable: true }
        category: { type: string, nullable: true }
        calendarId: { type: integer, format: int64, nullable: true }
        expirationTime: { type: string, format: date-time, nullable: true }
        disabled: { type: boolean }
        mileage: { type: number, format: double, nullable: true }
        attributes:
          type: object
          additionalProperties: {}
        ownerName: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DeviceInput:
      type: object
      required: [name, uniqueId]
      properties:
        name: { type: string }
        uniqueId: { type: string }
        phone: { type: string }
        model: { type: string }
        contact: { type: string }
        category: { type: string }
        calendarId: { type: integer, format: int64 }
        speedLimit: { type: number, format: double }
        disabled: { type: boolean }
        attributes:
          type: object
          additionalProperties: {}
...
paths:
...
  /api/devices:
    get:
      operationId: listDevices
      summary: List devices for the authenticated user
      tags: [devices]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createDevice
      summary: Create a device
      tags: [devices]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInput'
      responses:
        "201":
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/devices/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      operationId: getDevice
      summary: Get a device by ID
      tags: [devices]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "200":
          description: Device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateDevice
      summary: Update a device
      tags: [devices]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceInput'
      responses:
        "200":
          description: Updated device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteDevice
      summary: Delete a device
      tags: [devices]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "204":
          description: Deleted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
  /api/users/{id}/devices:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      operationId: adminListUserDevices
      summary: Admin — list all devices for a user
      tags: [admin]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "200":
          description: Device list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /api/users/{id}/devices/{deviceId}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
      - name: deviceId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    post:
      operationId: adminAssignDevice
      summary: Admin — assign a device to a user
      tags: [admin]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "204":
          description: Device assigned
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: adminUnassignDevice
      summary: Admin — unassign a device from a user
      tags: [admin]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "204":
          description: Device unassigned
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
  /api/admin/devices:
    get:
      operationId: adminListDevices
      summary: Admin — list all devices across all users
      tags: [admin]
      security:
        - cookieAuth: []
        - bearerAuth: []
        - xAuthToken: []
      responses:
        "200":
          description: All devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
...